Move OperationStore implementation into own module
