+++ p2panda-rs/src/test_utils/db/stores/mod.rs
@@ -1,3 +1,4 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 mod entry;
+mod operation;
